requirements.txt
# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Caching==2.3.0
Werkzeug==3.0.1
gunicorn==21.2.0
gevent==23.9.1

# AI APIs
anthropic==0.45.0
openai==1.54.3
google-generativeai==0.8.3

# Database
psycopg2-binary==2.9.9
psycogreen==1.0.2
pgvector==0.3.6

# Background Tasks
celery==5.4.0
redis==5.2.0

# File Processing
pypdf==3.17.4
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
Pillow==10.1.0

# Auth
authlib==1.3.0
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
certifi==2023.11.17
httpx==0.27.2

# Embedding & RAG
tiktoken==0.8.0
numpy==1.26.4
